+++ Runtime/Contracts/IChunkFactory.cs
@@ -2,6 +2,6 @@ namespace EliotByte.InfinityGen
 {
 	public interface IChunkFactory<TChunk, TDimension>
 	{
-		TChunk Create(TDimension position, float chunkSize, LayerRegistry<TDimension> layerRegistry);
+		TChunk Create(TDimension position, int chunkSize, LayerRegistry<TDimension> layerRegistry);
 	}
 }
